Organize files
Due to hard links and how Xcode saves files, the Source/ directory fell behind and was not up to date. Now, the relevant files lie within the source directory and Xcode will reference that directory when building the Example project. Therefore, it's important that the entire repository is downloaded in order to use the example project.